Fix DTO type mismatches found via stored procedure cross-verification
Cross-referenced the real stored procedure definitions (provided by the
user) against every field already sent from AhaClaimService.cs. Zero
critical name mismatches across 33 SP calls, but found three type
mismatches:

- ChronicKidneyDiseaseSection.Gfr/SerumCalcium/SerumPth: uspSaveCKD
  declares these as varchar(100), not decimal.
- PhysicalExaminationSection.Pulse/Breathing/BloodPressure1/BloodPressure2:
  uspSavePhysicalExamination declares these as smallint, not decimal.
- ScreeningSchedule2023Extras.UrineAlbuminResult/UrineCreatinineResult:
  uspSaveScreeningTest2023 declares these as varchar, not decimal.

Also verified EyesAndNeurologySection.RetinopathySeverity/
ProliferativeSeverity (sent as int? to both the pre-2023 SP, which
declares varchar, and the 2023+ SP, which declares int) matches legacy
behavior exactly -- AHADataAdapter.vb sends VerifyIntegerNull(...) for
both calls, relying on SQL Server's implicit int-to-varchar conversion.
No change needed there.

+++ Chopper/src/Chopper.Services/AhaClaims/PhysicalExaminationSection.cs
@@ -4,10 +4,10 @@ public sealed record PhysicalExaminationSection
 {
     public decimal? Temperature { get; init; }
     public string? TemperatureType { get; init; }
-    public decimal? Pulse { get; init; }
-    public decimal? Breathing { get; init; }
-    public decimal? BloodPressure1 { get; init; }
-    public decimal? BloodPressure2 { get; init; }
+    public int? Pulse { get; init; }
+    public int? Breathing { get; init; }
+    public int? BloodPressure1 { get; init; }
+    public int? BloodPressure2 { get; init; }
     public decimal? Height { get; init; }
     public string? HeightType { get; init; }
     public decimal? Weight { get; init; }
